feat: central snapshot helper
