+++ src/app/api/export/route.ts
@@ -1,10 +1,9 @@
 import { NextRequest } from "next/server";
 import { makeZip, type ZipFile } from "../../../lib/utils/zip";
 import { runWithFallback } from "../../../lib/providers/router";
-import { mkdir, readFile, writeFile } from "fs/promises";
-import path from "path";
 import crypto from "crypto";
 import { checkIdempotency } from "../../../lib/utils/idempotency";
+import { saveSnapshot } from "../../../lib/utils/snapshot";
 
 export const runtime = "nodejs";
 
@@ -47,43 +46,6 @@ function sha256Hex(data: Uint8Array | string) {
 
 function isoNow() { return new Date().toISOString(); }
 
-function tsFolder(d = new Date()) {
-  const pad = (n: number) => String(n).padStart(2, "0");
-  return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())}_${pad(d.getHours())}${pad(d.getMinutes())}`;
-}
-
-async function saveSnapshot(files: ZipFile[], target: string, lang: string, slug: string) {
-  const now = new Date();
-  const tsDir = tsFolder(now);
-  const timestamp = now.toISOString();
-  const entries: any[] = [];
-
-  for (const f of files) {
-    const data = typeof f.content === "string" ? Buffer.from(f.content) : Buffer.from(f.content);
-    const rel = path.join("snapshots", slug, tsDir, "paper", target, lang, f.path.replace(/^paper\//, ""));
-    const full = path.join(process.cwd(), "public", rel);
-    await mkdir(path.dirname(full), { recursive: true });
-    await writeFile(full, data);
-    entries.push({
-      path: rel.replace(/\\/g, "/"),
-      sha256: sha256Hex(data),
-      target,
-      lang,
-      slug,
-      timestamp
-    });
-  }
-
-  const manifestPath = path.join(process.cwd(), "public", "snapshots", "manifest.json");
-  let manifest: any[] = [];
-  try {
-    const existing = await readFile(manifestPath, "utf-8");
-    manifest = JSON.parse(existing);
-  } catch {}
-  manifest.push(...entries);
-  await mkdir(path.dirname(manifestPath), { recursive: true });
-  await writeFile(manifestPath, JSON.stringify(manifest, null, 2));
-}
 
 function buildTreeFromCompose(payload: any) {
   // EXPECTS:
